Add files via upload
The task of social media followers prediction is very valuable for every content creator and every business that relies on social media. So if you want to learn how to predict your social media followers for the next month, this article is for you. In this article, I will walk you through the task of social media followers prediction with machine learning using Python.